refactor(checkpoint): keep experimental APIs internal
Move the checkpoint implementation under nemo_gym._checkpoint until its contracts are ready for public support.
